Fix MSVC warnings about value truncations, webrtc/base/ edition.
BUG=chromium:81439
TEST=none
R=henrike@webrtc.org, marpan@google.com

Review URL: https://webrtc-codereview.appspot.com/20249004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@7143 4adac7df-926f-26a2-2b94-8c16560cd09d

+++ webrtc/base/socketadapters.cc
@@ -749,7 +749,7 @@ void AsyncSocksProxyServerSocket::HandleHello(ByteBuffer* request) {
   }
 }
 
-void AsyncSocksProxyServerSocket::SendHelloReply(int method) {
+void AsyncSocksProxyServerSocket::SendHelloReply(uint8 method) {
   ByteBuffer response;
   response.WriteUInt8(5);  // Socks Version
   response.WriteUInt8(method);  // Auth method
@@ -773,7 +773,7 @@ void AsyncSocksProxyServerSocket::HandleAuth(ByteBuffer* request) {
   state_ = SS_CONNECT;
 }
 
-void AsyncSocksProxyServerSocket::SendAuthReply(int result) {
+void AsyncSocksProxyServerSocket::SendAuthReply(uint8 result) {
   ByteBuffer response;
   response.WriteUInt8(1);  // Negotiation Version
   response.WriteUInt8(result);
